refactor: builder 관련 메서드 분리
