+++ app/model/trainer.py
@@ -1,5 +1,4 @@
 from lightfm.data import Dataset
-import random
 
 '''
 LightFM 모델 학습을 위한 데이터셋 구성, 상호작용 행렬 생성, 모델 학습을 수행하는 전체 파이프라인 함수들을 정의
@@ -20,28 +19,6 @@ def prepare_dataset(user_df, brand_df, user_feature_map, item_feature_map=None):
                 item_features=item_features)
 
     return dataset
-
-def build_interactions(dataset, interaction_df, user_brand_df, brand_df):
-    users_with_logs = set(interaction_df["user_id"])
-    all_users = set(user_brand_df["user_id"].unique())
-    users_without_logs = all_users - users_with_logs
-
-    dummy_interactions = []
-    for user_id in users_without_logs:
-        group = user_brand_df[user_brand_df["user_id"] == user_id]
-        recent = group[group["data_type"] == "RECENT"]["brand_id"].tolist()
-        interest = group[group["data_type"] == "INTEREST"]["brand_id"].tolist()
-
-        dummy_interactions += [(user_id, b, 2.0) for b in interest]
-        dummy_interactions += [(user_id, b, 3.0) for b in recent]
-
-        if not interest and not recent:
-            random.seed(user_id)
-            random_brand = random.choice(brand_df["brand_id"].tolist())
-            dummy_interactions.append((user_id, random_brand, 1.0))
-
-    real = list(zip(interaction_df["user_id"], interaction_df["brand_id"], interaction_df["weight"]))
-    return dataset.build_interactions(real + dummy_interactions)
 
 def train_model(interactions, weights, user_features, item_features):
     from lightfm import LightFM
